Update LICENSE
Dual License applied.  

Core program is released on GPL3, this is default unless otherwise confirmed in writing.  

If developers or contributors wish to incorporate into existing GUIs or adapt while not being required to share source code or release patent rights an alternative commercial licence can be applied.  The intent of this is to encourage open source collaboration but not remove the option of inclusion in larger proprietary works.