Initial commit of EVT for MODFLOW-USG
The following new parameters are introduced:

Output evt file switch
Can be either true or false
This parameter is optional. If not specified, it will be set to "false"
e.g output_evt: true

evt parameters (ievt, surf, exdp)
This parameter is optional
This parameter can accept a path to YAML or CSV
e.g evt_parameters: dummy_evt_params.csv

ievtcb
This parameter is an integer.
This parameter is optional. If not specified, it will be set to 50.
e.g ievtcb: 50

nevtopt
This parameter is an integer.
This parameter is optional. If not specified, it will be set to 2.
e.g nevtopt: 2